requirements.txt
streamlit
groq
python-dotenv
pillow
opencv-python
numpy
scikit-image
fpdf2==2.7.6
plotly
pandas 